Platform is not a submission method
Really, considering back to the Portal roots, the Platform is really just
a proxy or pass-through to existing submission methods and systems. By keeping
the details of additional information requested by agencies out of the
submission schema, we can ensure we stick to our goal of using the metadata to
describe the existing process. The inclusion of the additional fields is not so
much to facilitate dynamic forms on the Platform, but more about having agencies
self-describe their existing process which is useful to the FOIA big picture.