Add example metadata files (#80)
* Add example metadata files

* Update metadata file description

Include some of changes around describing submission methods. The description
doesn't quite match our examples as we're still figuring out if/how the 2015
scrapper metadata should be incorporated.

* Note on metadata schema origins

Focusing this metadata file on the current needs rather than making them an
iteration on the 2015 work.

* Update metadata schema description

Remove agency contact information for now as it's not relevant for submission.

* Clarify fields should be included

* Refer to post mail as paper

* Platform is not a submission method

Really, considering back to the Portal roots, the Platform is really just
a proxy or pass-through to existing submission methods and systems. By keeping
the details of additional information requested by agencies out of the
submission schema, we can ensure we stick to our goal of using the metadata to
describe the existing process. The inclusion of the additional fields is not so
much to facilitate dynamic forms on the Platform, but more about having agencies
self-describe their existing process which is useful to the FOIA big picture.

* Remove national_portal option

If folks want to change how electronic submissions work, they would alter their
submission methods as part of policy.