Simplify spec around additional form fields
Since we're not allowing the full uiSchema spec, it seems strange to include
them in a partial way. We can focus the metadata spec to describing specific
characteristics around the information the agency component wishes to collect.
We'll have a specific set of rules that transform these metadata fields into
JSON schema and uiSchema for react-jsonschema-form.